requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0

# Vector Store & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Text Processing
tiktoken>=0.5.0
rank-bm25>=0.2.2

# Web UI
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests 
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# Type hints
pydantic>=2.0.0
